requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
torch
pandas
beautifulsoup4
